transform is added to mesh struct

+++ src/pgl/mesh.c
@@ -10,7 +10,7 @@ const vertex_t triangle_vertices[3] = {
 	{{ 0.0f,   PGLM_1_SQRT3f,  0.0f}, {0.0f, 1.0f, 0.0f}},
 };
 
-const uint32_t triangle_indices[3] = {
+const uint16_t triangle_indices[3] = {
     0, 1, 2,
 };
 
@@ -24,7 +24,7 @@ const vertex_t square_vertices[4] = {
 	{{-0.5f,  0.5f,  0.0f}, {0.0f, 0.0f, 1.0f}},
 };
 
-const uint32_t square_indices[6] = {
+const uint16_t square_indices[6] = {
     0, 1, 2,
 	0, 2, 3,
 };
@@ -43,7 +43,7 @@ const vertex_t cube_vertices[8] = {
 	{{-0.5f,  0.5f,  0.5f}, {1.0f, 0.0f, 0.0f}},
 };
 
-const uint32_t cube_indices[36] = {
+const uint16_t cube_indices[36] = {
 	0, 1, 2,
 	0, 2, 3,
 	4, 5, 6,
@@ -68,7 +68,7 @@ const vertex_t triangle_pyramid_vertices[4] = {
 	{{ 0.0f, PGLM_SQRT2f/PGLM_SQRT3f, 		     0.0f}, {0.0f, 1.0f, 0.0f}},
 };
 
-const uint32_t triangle_pyramid_indices[12] = {
+const uint16_t triangle_pyramid_indices[12] = {
 	0, 2, 1,
 	0, 1, 3,
 	1, 2, 3,
@@ -86,7 +86,7 @@ const vertex_t square_pyramid_vertices[5] = {
 	{{ 0.0f, 0.8f,  0.0f}, {0.0f, 1.0f, 1.0f}}, 
 };
 
-const uint32_t square_pyramid_indices[18] = {
+const uint16_t square_pyramid_indices[18] = {
 	0, 1, 2,
 	0, 2, 3,
 	4, 0, 1,
